vertex attrib struct
