+++ oglw/gl/vertexAttribute.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+#include "glTypes.h"
+
+struct VertexAttrib {
+    std::string name;
+    GLint size;
+    GLenum type;
+    GLboolean normalized;
+    GLvoid* offset;
+};
